Documentation of registration and extrapolation
